30-97382-webapp-habit-tracker/code/backend/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
tortoise-orm==0.17.8
aerich==0.5.5
asyncpg==0.23.0
aiomysql==0.0.22
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5